finsentinel-backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
scikit-learn
xgboost
imbalanced-learn
pandas
numpy
joblib
python-dotenv
uuid
shap
flask-limiter
kafka-python
networkx
flask-jwt-extended
gunicorn
pytest